+++ Infinity/GameMode/Notification/StaticsManager.tscn
@@ -1,0 +1,8 @@
+[gd_scene load_steps=3 format=3 uid="uid://cxq3307e0nngm"]
+
+[ext_resource type="Script" path="res://Infinity/GameMode/Notification/StaticsManager.cs" id="1_5hvg3"]
+[ext_resource type="PackedScene" uid="uid://5ck4fqkx4pvq" path="res://Infinity/GameMode/Notification/DamageNumberHud.tscn" id="2_rkwjy"]
+
+[node name="NotificationManager" type="Node"]
+script = ExtResource("1_5hvg3")
+_damageNumberScene = ExtResource("2_rkwjy")
