+++ Infinity/Infinity.tscn
@@ -1,4 +1,4 @@
-[gd_scene load_steps=43 format=3 uid="uid://c8nr5w2kj67dx"]
+[gd_scene load_steps=42 format=3 uid="uid://c8nr5w2kj67dx"]
 
 [ext_resource type="Script" path="res://Infinity/GameMode/Main.cs" id="1_tk8vx"]
 [ext_resource type="Script" path="res://Infinity/Datebase/Shop/ShopConfig.cs" id="2_2pohd"]
@@ -20,8 +20,7 @@
 [ext_resource type="PackedScene" uid="uid://bali3cj5bhs2q" path="res://Infinity/HUDs/InGameHUD.tscn" id="11_3i81h"]
 [ext_resource type="PackedScene" uid="uid://cwjfqs0u6t7wi" path="res://Infinity/HUDs/TooltipToast.tscn" id="12_42duw"]
 [ext_resource type="PackedScene" uid="uid://bfke7q0b0owyb" path="res://Infinity/Player/MeMe.tscn" id="12_e0mmc"]
-[ext_resource type="Script" path="res://Infinity/GameMode/Notification/NotificationManager.cs" id="13_mqafi"]
-[ext_resource type="PackedScene" uid="uid://5ck4fqkx4pvq" path="res://Infinity/GameMode/Notification/DamageNumberHud.tscn" id="14_mtewq"]
+[ext_resource type="PackedScene" uid="uid://cxq3307e0nngm" path="res://Infinity/GameMode/Notification/StaticsManager.tscn" id="13_d8606"]
 [ext_resource type="PackedScene" uid="uid://cfoh2vs087f48" path="res://PoseMenu/PoseMenu.tscn" id="16_ngbfj"]
 [ext_resource type="PackedScene" uid="uid://dh7nt37deh6b1" path="res://Common/Debugging/DebugHUD.tscn" id="16_qck06"]
 [ext_resource type="Script" path="res://Infinity/GameMode/SoundManager.cs" id="23_nx00x"]
@@ -98,23 +97,20 @@ size = Vector2(100, 1200)
 [sub_resource type="RectangleShape2D" id="RectangleShape2D_rh687"]
 size = Vector2(1200, 100)
 
-[node name="Main" type="Node" node_paths=PackedStringArray("_playerController", "_playerPawn", "_enemySpawner")]
+[node name="Main" type="Node" node_paths=PackedStringArray("_playerPawn", "_enemySpawner")]
 script = ExtResource("1_tk8vx")
 _gameSettings = SubResource("Resource_pdcir")
-_playerController = NodePath("PlayerController")
-_playerPawn = NodePath("MeMe (Player)")
+_playerPawn = NodePath("Player")
 _enemySpawner = NodePath("EnemySpawner")
 
-[node name="NotificationManager" type="Node" parent="." node_paths=PackedStringArray("_damageNumberContainer")]
-script = ExtResource("13_mqafi")
-_damageNumberScene = ExtResource("14_mtewq")
-_damageNumberContainer = NodePath("../HUDs/DamageHud (Layer 3)")
+[node name="StaticsManager" parent="." node_paths=PackedStringArray("_customDamageNumberContainer") instance=ExtResource("13_d8606")]
+_customDamageNumberContainer = NodePath("../HUDs/DamageHud (Layer 3)")
 
 [node name="PlayerController" type="Node" parent="."]
 process_mode = 1
 script = ExtResource("6_m3lsh")
 
-[node name="MeMe (Player)" parent="." instance=ExtResource("12_e0mmc")]
+[node name="Player" parent="." instance=ExtResource("12_e0mmc")]
 
 [node name="EnemySpawner" type="Node2D" parent="." node_paths=PackedStringArray("_enemySpawnRoot", "_spawnPath")]
 process_mode = 1
@@ -176,9 +172,6 @@ visible = false
 [node name="DamageHud (Layer 3)" type="CanvasLayer" parent="HUDs"]
 layer = 3
 follow_viewport_enabled = true
-
-[node name="DamageNumberHud" parent="HUDs/DamageHud (Layer 3)" instance=ExtResource("14_mtewq")]
-position = Vector2(314, 181)
 
 [node name="BattleResult HUD (Layer 2)" parent="HUDs" instance=ExtResource("8_gtxqj")]
